rust-rewrite: mcp refactor + verbose log
