+++ .vscode/mcp.json
@@ -2,6 +2,9 @@
   "servers": {
     "cfd-rs-memory": {
       "type": "stdio",
+      "env": {
+        "MCP_LOG": "brief"
+      },
       "command": "cargo",
       "args": [
         "run",
